Adding direct URL download functionality
This change contains an implementation of JIRA task ABC-123 which allows a user
to retrieve a file from the web and store it in their file store. On visiting
the appropriate URL, a user is presented with a form where they fill in the URL
of the file and the title under which they wish to store their copy. The backend
then downloads that file and saves it to the file store.

The following remains before this feature can be considered complete:
 - The file metadata must be stored to the database; this will be added once task ABC-124 is complete.
 - The JSP pages will be formatted to match the UX guide in a later commit.
 - The tests for the HttpUtils class remain to be implemented.